+++ config/routes.rb
@@ -29,7 +29,7 @@ Rails.application.routes.draw do
     resources :activities, only: [], concerns: [:transactionable, :budgetable] do
       resources :steps, controller: "activity_forms"
       resources :extending_organisations, only: [:edit, :update]
-      resources :implementing_organisations, only: [:new, :create]
+      resources :implementing_organisations, only: [:new, :create, :edit, :update]
     end
   end