+++ CHANGELOG.md
@@ -51,3 +51,4 @@
 - Users can edit a budget
 - Users can add currency information to budgets and see the currency in the budget XML
 - Users can report an activities implementing organisation
+- Users can edit a reported implementing organisation